requirements.txt
fastapi
uvicorn
langgraph
pydantic
pydantic-settings
pymongo
psycopg2-binary
# Add MCP/A2A SDK stubs here if they were real packages
